types

+++ bar_extract.c
@@ -990,11 +990,11 @@ int bar_extract(const struct logcb *log, const char *archive, struct jlhead *fil
 					strftime(buf, sizeof(buf), "%F %T", &tm);
 					if(conf->pkginfo)
 						printf("%s\t%llu\t%s\t%s\n",
-						       cpio.mode, cpio.c_filesize, filemd5?filemd5:"-", cpio.name);
+						       cpio.mode, (long long unsigned int) cpio.c_filesize, filemd5?filemd5:"-", cpio.name);
 					else
 						printf("%-8o %4d %6s %6s %9llu %9s %s %s\n",
 						       cpio.c_mode, cpio.c_nlink, fileuser, filegroup,
-						       cpio.c_filesize, buf, cpio.name, fileflags?fileflags:"");
+						       (long long unsigned int) cpio.c_filesize, buf, cpio.name, fileflags?fileflags:"");
 				} else
 					printf("%s\n", cpio.name);
 			}
